pyproject.toml
[project]
name = "aws-hub-spoke-ecs"
version = "0.1.0"
description = "Hub-spoke ECS deployment with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "boto3>=1.42.23",
    "python-dotenv>=1.2.1",
]
